+++ cypress/support/commands.js
@@ -24,8 +24,8 @@ Cypress.Commands.add('getPickupPointItem', () => {
 })
 
 Cypress.Commands.add('visitStore', () => {
-  cy.intercept('**/rc.vtex.com.br/api/events').as('events')
-  cy.visit('/stores')
+  cy.intercept('**/event-api/v1/productusqa/event').as('events')
+  cy.visit('/api/io/stores')
   cy.wait('@events')
 
   cy.get(selectors.ProfileLabel, { timeout: 20000 })