+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
++ ### Fixed
++ - removing the added holiday day, and included the locale sent in the intl.
 
 ## [0.10.15] - 2023-01-11